# backend/storage/api-docs-admin/api-docs-admin.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Next Starter API — Admin",
        "description": "Documentação da área **Admin** (equipe interna / back-office) do projeto Laravel Next Starter.\n\nEsta documentação é restrita: o acesso à interface Swagger UI e ao JSON da spec exige autenticação HTTP Basic própria da documentação (usuário/senha definidos em `.env`), **independente** do login da API.\n\nApós acessar a documentação, os endpoints em si continuam exigindo o Bearer JWT obtido em `POST /api/admin/login`.",
        "contact": {
            "name": "Equipe de Desenvolvimento"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API (relativo ao host onde o Swagger UI está sendo servido)"
        }
    ],
    "paths": {
...
        "/admin/mensagens": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Listar mensagens enviadas",
                "operationId": "0369f09daef8284df1fe0430c0e6dd30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "titulo",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "maxLength": 120
                        }
                    },
                    {
                        "name": "origem",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "sistema",
                                "admin"
                            ]
                        }
                    },
                    {
                        "name": "por_pagina",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de mensagens.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AdminMensagem"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Cadastrar/enviar mensagem",
                "description": "Cria e envia uma mensagem/notificação, com direcionamento configurável: geral (todos os usuários), por entidade (Admin ou Private), por grupo empresa, ou para um usuário específico.",
                "operationId": "8245b22ddc85c39bddd7dd4db6c89e51",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "titulo",
                                    "conteudo",
                                    "direcionamento"
                                ],
                                "properties": {
                                    "titulo": {
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "conteudo": {
                                        "type": "string"
                                    },
                                    "direcionamento": {
                                        "required": [
                                            "tipo"
                                        ],
                                        "properties": {
                                            "tipo": {
                                                "type": "string",
                                                "enum": [
                                                    "geral",
                                                    "entidade",
                                                    "grupo_empresa",
                                                    "usuario"
                                                ]
                                            },
                                            "entidade_tipo": {
                                                "description": "Obrigatório quando tipo = entidade.",
                                                "type": "string",
                                                "enum": [
                                                    "admin",
                                                    "private"
                                                ]
                                            },
                                            "grupo_empresa_id": {
                                                "description": "Obrigatório quando tipo = grupo_empresa.",
                                                "type": "string",
                                                "format": "uuid"
                                            },
                                            "usuario_id": {
                                                "description": "Obrigatório quando tipo = usuario.",
                                                "type": "string",
                                                "format": "uuid"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mensagem cadastrada e enviada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AdminMensagem"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/mensagens/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Visualizar mensagem",
                "operationId": "a7d5cda3da870557f81357b2289303b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagem encontrada.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AdminMensagem"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/mensagens/usuarios": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Admin — Buscar usuários para destinatário de mensagem",
                "description": "Busca usuários de qualquer entidade (Admin ou Private) por nome, usada exclusivamente para selecionar o destinatário individual no formulário de envio de mensagem.",
                "operationId": "35d186a26e990266be7505e2b175db93",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "por_pagina",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 50,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de usuários.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid"
                                                    },
                                                    "nome": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminMensagem": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "titulo": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "conteudo": {
                        "type": "string"
                    },
                    "origem": {
                        "type": "string",
                        "enum": [
                            "sistema",
                            "admin"
                        ]
                    },
                    "remetente": {
                        "properties": {
                            "id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "nome": {
                                "type": "string"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    },
                    "direcionamento": {
                        "properties": {
                            "tipo": {
                                "type": "string",
                                "enum": [
                                    "geral",
                                    "entidade",
                                    "grupo_empresa",
                                    "usuario"
                                ]
                            },
                            "entidade_tipo": {
                                "type": "string",
                                "enum": [
                                    "admin",
                                    "private"
                                ],
                                "nullable": true
                            },
                            "grupo_empresa_id": {
                                "type": "string",
                                "format": "uuid",
                                "nullable": true
                            },
                            "grupo_empresa_nome": {
                                "type": "string",
                                "nullable": true
                            },
                            "usuario_id": {
                                "type": "string",
                                "format": "uuid",
                                "nullable": true
                            },
                            "usuario_nome": {
                                "type": "string",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "total_destinatarios": {
                        "type": "integer",
                        "nullable": true
                    },
                    "total_lidos": {
                        "type": "integer",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "description": "Meta de paginação padrão do Laravel (LengthAwarePaginator serializado\npor um ResourceCollection), usada em todos os endpoints \"listar\".",
                        "type": "string",
                        "format": "uri",
                        "example": "/api/admin/empresas?page=1",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "format": "uri",
                        "example": "/api/admin/empresas?page=5",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "format": "uri",
                        "example": null,
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "format": "uri",
                        "example": "/api/admin/empresas?page=2",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "path": {
                        "type": "string",
                        "example": "/api/admin/empresas"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 68
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "errors": {
                        "description": "Formato de erro de validação (422), produzido pelo handler de exceções\nglobal (bootstrap/app.php) a partir de uma ValidationException.",
                        "type": "object",
                        "example": {
                            "campo": [
                                "A mensagem de validação correspondente."
                            ]
                        },
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "BusinessErrorResponse": {
                "properties": {
                    "errors": {
                        "properties": {
                            "business": {
                                "description": "Formato padrão de erro de negócio/HTTP (400/401/403/404/500), produzido\npelo mesmo handler global para BusinessException, AuthorizationException,\nModelNotFoundException, HttpExceptionInterface, etc.",
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "Descrição do erro de negócio."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "ValidationError": {
                "description": "Erro de validação dos campos enviados.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ValidationErrorResponse"
                        }
                    }
                }
            },
...
            "Unauthorized": {
                "description": "Não autenticado: token ausente, inválido, expirado, ou sessão associada ao token não está mais ativa.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Autenticado, porém sem permissão para executar esta ação.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Registro não encontrado.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/BusinessErrorResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticação via **JSON Web Token (JWT)**.\n\n1. Faça login em `POST /api/login` (Private) ou `POST /api/admin/login` (Admin).\n2. Envie o token retornado no header `Authorization: Bearer {token}` nas demais requisições.\nO token carrega um `session_id`, validado a cada requisição contra a sessão ativa do usuário (tabela de sessões).",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}